SW-1749 Split SearchServiceTest
- Create abstract base class for SearchServiceTest
  - Inherits DatabaseTest
  - Implements RunAsUser
  - Contains helper vals/funs
- Divide tests by purpose
